+++ .copier-answers.yml
@@ -1,0 +1,38 @@
+# Changes here will be overwritten by Copier; NEVER EDIT MANUALLY
+_commit: 02b14ca
+_src_path: git+https://github.com/twsl/python-project-template
+author_email: twsl@users.noreply.github.com
+author_username: twsl
+copyright_holder: twsl
+copyright_holder_email: twsl@users.noreply.github.com
+copyright_license: MIT
+copyright_year: 2024
+custom_install: true
+docs_library: mkdocs
+github_runner:
+- ubuntu-latest
+github_runner_python_version:
+- '3.8'
+- '3.9'
+- '3.10'
+- '3.11'
+- '3.12'
+include_copilot: true
+include_databricks: false
+include_docs: true
+include_notebooks: true
+include_sample_code: false
+line_ending: lf
+primary_branch: main
+project_description: The spiritual successor to knockknock for PyTorch Lightning,
+    get notified when your training ends
+project_name: whos_there
+python_package_command_line_name: whos-there
+python_package_distribution_name: whos-there
+python_package_import_name: whos_there
+python_version: '3.11'
+repository_name: whos-there
+repository_namespace: twsl
+repository_provider: github.com
+self_signed: false
+use_precommit: true
