+++ .github/CODEOWNERS
@@ -1,0 +1,11 @@
+# Each line is a file pattern followed by one or more owners.
+# These owners will be the default owners for everything in
+# the repo. Unless a later match takes precedence,
+# @global-owner1 and @global-owner2 will be requested for
+# review when someone opens a pull request.
+
+* @twsl
+
+/.github/ @twsl
+/src/ @twsl
+/tests/ @twsl
