+++ .github/ISSUE_TEMPLATE/1_bug_report.yaml
@@ -1,0 +1,95 @@
+name: Bug Report
+description: File a bug report
+title: "[Bug]: "
+labels: ["bug", "triage"]
+assignees:
+  - twsl
+body:
+  - type: markdown
+    attributes:
+      value: |
+        Thanks for taking the time to contibute to whos_there!
+
+  - type: textarea
+    attributes:
+      label: Bug description
+      description: A description of the 🐛bug🐛.
+      placeholder: |
+        A clear and concise description of what the bug is.
+
+        Include steps to reproduce, the expected behaviour, and the actual behaviour.
+
+        ```
+        The error message you got, with the full traceback if available.
+        ```
+    validations:
+      required: true
+
+  - type: input
+    id: version
+    attributes:
+      label: Software version
+      description: What version of whos_there are you running?
+      placeholder: ex. latest
+    validations:
+      required: false
+
+  - type: markdown
+    attributes:
+      value: "**Note: The rest of this form is optional, but filling it out may help us to provide better support.**"
+
+  - type: textarea
+    attributes:
+      label: How to reproduce the bug
+      description: >
+        Provide steps and example code here.
+      placeholder: |
+        ```python
+        # Sample code to reproduce the problem
+        ```
+      render: python
+    validations:
+      required: false
+
+  - type: textarea
+    attributes:
+      label: Error messages and logs
+      description: >
+        Provide any error messages and/or logs
+      placeholder: "# Copy the complete error messages and logs"
+      value: |
+        ```
+        # Error messages and logs here please
+        ```
+    validations:
+      required: false
+
+  - type: textarea
+    attributes:
+      label: Environment
+      description: |
+        Please provide details about your environment.
+      value: |
+        <details>
+          <summary>Current environment</summary>
+
+        ```
+        #- Python version (e.g., 3.9):
+        #- OS (e.g., Linux):
+        #- How you installed Lightning(`conda`, `pip`, source):
+        ```
+
+        </details>
+    validations:
+      required: false
+
+  - type: textarea
+    attributes:
+      label: More info
+      description: Add any other info about the issue here.
+    validations:
+      required: false
+
+  - type: markdown
+    attributes:
+      value: "**Happy engineering!**"
