+++ .github/ISSUE_TEMPLATE/2_feature_request.yaml
@@ -1,0 +1,33 @@
+name: Feature request
+description: Propose a feature for this project
+labels: ["triage", "feature"]
+body:
+  - type: textarea
+    attributes:
+      label: Description & Motivation
+      description: A clear and concise description of the feature proposal.
+      placeholder: |
+        Please outline the motivation for the proposal.
+        Is your feature request related to a problem? e.g., I'm always frustrated when [...].
+        If this is related to another GitHub issue, please link it here.
+
+  - type: textarea
+    attributes:
+      label: Pitch
+      description: A clear and concise description of what you want to happen.
+    validations:
+      required: false
+
+  - type: textarea
+    attributes:
+      label: Alternatives
+      description: A clear and concise description of any alternative solutions or features you've considered, if any.
+    validations:
+      required: false
+
+  - type: textarea
+    attributes:
+      label: Additional context
+      description: Add any other context or screenshots about the feature request here.
+    validations:
+      required: false
