+++ .github/ISSUE_TEMPLATE/4_documentation.yaml
@@ -1,0 +1,16 @@
+name: Typos and doc fixes
+description: Tell us about how to improve our documentation
+labels: ["triage", "docs"]
+body:
+  - type: textarea
+    attributes:
+      label: 📚 Documentation
+      description: A description of what you want to improve..
+      placeholder: |
+        For typos and doc fixes, please go ahead and:
+
+        - For a simlpe typo or fix, please send directly a PR (no need to create an issue)
+        - If you are not sure about the proper solution, please describe here your finding...
+
+    validations:
+      required: true
