+++ .github/ISSUE_TEMPLATE/config.yml
@@ -3,3 +3,5 @@ contact_links:
   - name: Ask a Question
     url: https://github.com/twsl/whos-there/discussions/new
     about: Ask and answer questions
+  - name: Read the documentation
+    url: https://pages.github.com/twsl/whos-there/
