+++ .github/dependabot.yml
@@ -1,31 +1,33 @@
+---
 # Configuration: https://dependabot.com/docs/config-file/
 # Docs: https://docs.github.com/en/github/administering-a-repository/keeping-your-dependencies-updated-automatically
 
 version: 2
 
 updates:
-  - package-ecosystem: "pip"
+  - package-ecosystem: "gitsubmodule"
     directory: "/"
     schedule:
-      interval: "daily"
-    allow:
-      - dependency-type: "production"
-      - dependency-type: "development"
-    #commit-message:
-      #prefix: ":arrow_up:"
-    open-pull-requests-limit: 50
-
-  - package-ecosystem: "github-actions"
-    directory: "/"
-    schedule:
-      interval: "daily"
-    allow:
-      - dependency-type: "all"
-    #commit-message:
-      #prefix: ":arrow_up:"
-    open-pull-requests-limit: 50
+      interval: daily
 
   - package-ecosystem: "devcontainers"
     directory: "/"
     schedule:
       interval: daily
+
+  - package-ecosystem: github-actions
+    directory: /
+    schedule:
+      interval: daily
+    allow:
+      - dependency-type: all
+    open-pull-requests-limit: 50
+
+  - package-ecosystem: pip
+    directory: /
+    schedule:
+      interval: daily
+    allow:
+      - dependency-type: production
+      - dependency-type: development
+    open-pull-requests-limit: 50
