+++ .github/release-drafter.yaml
@@ -1,0 +1,60 @@
+---
+# Release drafter configuration https://github.com/release-drafter/release-drafter#configuration
+# Emojis were chosen to match the https://gitmoji.carloscuesta.me/
+
+name-template: v$RESOLVED_VERSION
+tag-template: v$RESOLVED_VERSION
+
+change-template: "- $TITLE @$AUTHOR (#$NUMBER)"
+change-title-escapes: '\<*_&' # You can add # and @ to disable mentions
+
+categories:
+  - title: ":rocket: Features"
+    labels:
+      - feature
+      - enhancement
+  - title: ":wrench: Fixes & Refactoring"
+    labels:
+      - bug
+      - bugfix
+      - fix
+      - refactoring
+  - title: ":package: Build System & CI/CD"
+    labels:
+      - build
+      - ci
+      - testing
+  - title: ":boom: Breaking Changes"
+    labels:
+      - breaking
+  - title: ":pencil: Documentation"
+    labels:
+      - documentation
+      - docs
+  - title: ":broom: Maintenance"
+    labels:
+      - maintenance
+      - refactoring
+      - cosmetic
+      - chore
+  - title: ":arrow_up: Dependencies updates"
+    labels:
+      - dependencies
+
+version-resolver:
+  major:
+    labels:
+      - major
+  minor:
+    labels:
+      - minor
+  patch:
+    labels:
+      - patch
+  default: patch
+
+template: |
+  ## What's Changed
+  $CHANGES
+  ## :busts_in_silhouette: List of contributors
+  $CONTRIBUTORS
