--- .github/workflows/build.yml
@@ -1,94 +1,0 @@
-name: Build
-
-on: [push, pull_request]
-
-jobs:
-
-  build-and-test:
-    name: "Build and Test"
-    runs-on:  ${{ matrix.os }}
-    defaults:
-      run:
-        shell: bash
-    strategy:
-      matrix:
-        os: [ubuntu-latest, macos-latest, "windows-latest"]
-        python-version: ["3.8", "3.9", "3.10", "3.11"]
-
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v4
-
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install Poetry
-      uses: snok/install-poetry@v1
-      with:
-        virtualenvs-create: true
-        virtualenvs-in-project: true
-
-    - name: Set up cache
-      uses: actions/cache@v4
-      id: cached-poetry-dependencies
-      with:
-        path: .venv
-        key: venv-${{ matrix.os }}-${{ matrix.python-version }}-${{ hashFiles('pyproject.toml') }}-${{ hashFiles('poetry.lock') }}
-
-    - name: Install dependencies
-      if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
-      run: poetry install --no-interaction
-
-    - name: PyTorch version
-      run: |
-        poetry run python -c "import torch; print(torch.__version__)"
-
-    - name: Run ruff
-      uses: chartboost/ruff-action@v1
-
-    - name: Install plugin
-      run: pip install pytest-github-actions-annotate-failures
-
-    - name: Run tests
-      run: |
-        poetry run pytest -c pyproject.toml --cov-report=xml --cov-report=html --junitxml=junit/test-results-${{ matrix.python-version }}.xml --cov=whos_there
-
-    - name: Run safety checks
-      run: |
-        poetry check
-        poetry run safety check --full-report
-        poetry run bandit -ll --recursive hooks
-      continue-on-error: true
-
-    - name: Upload test results
-      uses: actions/upload-artifact@v4
-      with:
-        name: pytest-results-${{ matrix.python-version }}
-        path: junit/test-results-${{ matrix.python-version }}.xml
-      if: ${{ always() }}
-
-  publish:
-    name: "Publish Unit Tests Results"
-    needs: build-and-test
-    runs-on: ubuntu-latest
-    # the build-and-test job might be skipped, we don't need to run this job then
-    if: >
-      always() &&
-      github.event.workflow_run.conclusion != 'skipped' && (
-        github.event.sender.login != 'dependabot[bot]' ||
-        github.event.workflow_run.head_repository.full_name != github.repository
-      )
-
-    steps:
-      - name: Download Artifacts
-        uses: actions/download-artifact@v4
-        with:
-          path: artifacts
-
-      - name: Publish Unit Test Results
-        uses: EnricoMi/publish-unit-test-result-action@v2
-        if: github.event.sender.login != 'dependabot[bot]'
-        with:
-          files: artifacts/**/*.xml
