+++ .github/workflows/docs.yaml
@@ -1,0 +1,103 @@
+---
+name: Documentation
+
+on:
+  push:
+  pull_request:
+  workflow_dispatch:
+
+# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
+permissions:
+  contents: read
+  pages: write
+  id-token: write
+
+# Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
+# However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
+concurrency:
+  group: "pages"
+  cancel-in-progress: false
+
+env:
+  ARTIFACT: docs
+
+
+jobs:
+  build-docs:
+    name: Build docs
+    runs-on: ['ubuntu-latest']
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Install Poetry
+        uses: snok/install-poetry@v1
+        with:
+          virtualenvs-create: true
+          virtualenvs-in-project: true
+
+      - name: Set up cache
+        uses: actions/cache@v4
+        id: cached-poetry-dependencies
+        with:
+          path: .venv
+          key: venv-Linux-3.11-${{ hashFiles('poetry.lock') }}
+
+      - name: Sphinx
+        run: sudo apt install -y python3-sphinx
+
+      - name: Install dependencies
+        if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+        run: poetry install --no-interaction
+
+      - name: Build dependencies
+        run: poetry build
+
+      - name: Build MkDocs docs
+        run: poetry run mkdocs build -f ./docs/mkdocs.yml -d ./docs/_build/
+
+      - name: Upload build artifacts
+        uses: actions/upload-artifact@v3
+        with:
+          name: ${{ env.ARTIFACT }}
+          path: ./docs/build/
+        if: ${{ always() }}
+
+  deploy-docs:
+    name: Deploy docs
+    needs: build-docs
+    permissions:
+      pages: write
+      id-token: write
+    runs-on: ['ubuntu-latest']
+    if: github.ref == 'refs/heads/main'
+    environment:
+      name: github-pages
+      url: ${{ steps.deployment.outputs.page_url }}
+    steps:
+      - name: Setup Pages
+        uses: actions/configure-pages@v4
+
+      - name: Download Artifacts
+        uses: actions/download-artifact@v3
+        with:
+          path: artifacts
+
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@v2 # cant update yet
+        with:
+          artifact_name: "github-pages"
+          path: artifacts/${{ env.ARTIFACT }}/
+
+      - name: Deploy to GitHub Pages
+        id: deployment
+        uses: actions/deploy-pages@v3 # cant update yet
+        with:
+          artifact_name: "github-pages"
