+++ .github/workflows/pre-commit.yaml
@@ -1,0 +1,30 @@
+---
+name: pre-commit
+
+on:
+  pull_request:
+  push:
+    branches: [main]
+
+
+jobs:
+  pre-commit:
+    name: pre-commit
+    strategy:
+      fail-fast: true
+
+      matrix:
+        python-version: ['3.8', '3.9', '3.10', '3.11', '3.12']
+    runs-on: ['ubuntu-latest']
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Run pre-commit
+        uses: pre-commit/action@v3.0.1
