+++ .pre-commit-config.yaml
@@ -1,70 +1,74 @@
 default_language_version:
-  python: python3.10
+  python: python3
 
 default_stages: [commit, push]
 
 ci:
   autofix_prs: true
-  autoupdate_commit_msg: '[pre-commit.ci] pre-commit suggestions'
+  autoupdate_commit_msg: "[pre-commit.ci] pre-commit suggestions"
   autoupdate_schedule: weekly
 
 repos:
   - repo: https://github.com/python-poetry/poetry
-    rev: 1.5.1
+    rev: 1.8.0
     hooks:
       - id: poetry-check
       - id: poetry-lock
 
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.4.0
+    rev: v4.6.0
     hooks:
+      - id: check-merge-conflict
+        args: [--assume-in-merge]
       - id: end-of-file-fixer
       - id: trailing-whitespace
       - id: check-json
+        exclude: devcontainer.json
       - id: check-yaml
+        args: [--unsafe]
       - id: check-toml
       - id: check-docstring-first
       - id: check-executables-have-shebangs
       - id: check-case-conflict
       - id: check-added-large-files
-        args: ["--maxkb=200", "--enforce-all"]
+        args: ["--maxkb=500", "--enforce-all"]
       - id: detect-private-key
 
-  - repo: https://github.com/asottile/pyupgrade
-    rev: v3.7.0
+  - repo: https://github.com/PyCQA/bandit
+    rev: 1.7.8
     hooks:
-      - id: pyupgrade
-        args: [--py37-plus]
-        name: Upgrade code
+      - id: bandit
+        args: ["-c", "pyproject.toml"]
+        additional_dependencies: ["bandit[toml]"]
 
-  - repo: https://github.com/psf/black
-    rev: 23.3.0
+  - repo: https://github.com/Lucas-C/pre-commit-hooks-safety
+    rev: v1.3.3
     hooks:
-      - id: black
-        name: Format code
-        args: ["--config=pyproject.toml", "-v"]
-
-  - repo: https://github.com/PyCQA/isort
-    rev: 5.12.0
-    hooks:
-      - id: isort
-        name: Format imports
-        args: ["--settings-path=pyproject.toml", "-v"]
-
-  - repo: https://github.com/PyCQA/flake8
-    rev: 6.0.0
-    hooks:
-      - id: flake8
-        name: Check PEP8
-        args: ["--ignore=E24,W504,F401,E203,W503", "--max-line-length=120",]
+      - id: python-safety-dependencies-check
+        files: pyproject.toml
+        args: ["--groups=dev,test"]
 
   - repo: https://github.com/kynan/nbstripout
-    rev: 0.6.1
+    rev: 0.7.1
     hooks:
       - id: nbstripout
 
   - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: v0.0.275
+    rev: v0.4.1
     hooks:
       - id: ruff
         args: ["--fix", "-v"]
+
+  - repo: https://github.com/RobertCraigie/pyright-python
+    rev: v1.1.359
+    hooks:
+      - id: pyright
+        entry: poetry run pyright -p .
+        language: system
+        #pass_filenames: false
+        types: [python]
+
+  - repo: https://github.com/gitleaks/gitleaks
+    rev: v8.18.2
+    hooks:
+      - id: gitleaks
