+++ docs/docs/notebooks/index.md
@@ -1,0 +1,3 @@
+# Example Notebooks
+
+- [Boring model](the-boring-model.ipynb)
