+++ docs/docs/styles/extra.css
@@ -1,0 +1,3 @@
+.md-typeset .md-code__content {
+    display: block !important;
+}
