+++ pyproject.toml
@@ -2,24 +2,22 @@
 requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
 build-backend = "poetry_dynamic_versioning.backend"
 
-[tool.poetry-dynamic-versioning]
-enable = true
-vcs = "git"
-style = "semver"
-pattern = "default-unprefixed"
+[project]
+name = "whos-there"
+requires-python = ">=3.8"
 
+# TODO: https://github.com/python-poetry/poetry/issues/3332
 [tool.poetry]
 name = "whos-there"
-version = "0.1.5-post.110+1e86ce1"
+version = "0.0.0"
 description = "The spiritual successor to knockknock for PyTorch Lightning, get notified when your training ends"
 authors = ["twsl <45483159+twsl@users.noreply.github.com>"]
 maintainers = ["twsl <45483159+twsl@users.noreply.github.com>"]
 license = "MIT"
 readme = "README.md"
-packages = [{include = "whos_there", from = "src"}]
-homepage = "https://github.com/twsl/whos-there"
 repository = "https://github.com/twsl/whos-there"
-documentation = "https://github.com/twsl/whos-there"
+homepage = "https://pages.github.com/twsl/whos-there"
+packages = [ { include = "whos_there", from = "src" } ]
 keywords = [
   "python",
   "pytorch",
@@ -33,38 +31,81 @@ classifiers = [
     "Topic :: Software Development"
 ]
 
+[tool.poetry.urls]
+issues = "https://github.com/twsl/whos-there/issues"
+
+
 [tool.poetry.dependencies]
-python = "^3.8.1"
-pytorch-lightning = ">=2.0,<3.0"
-python-telegram-bot = ">=20,<22"
+python = "^3.11"
+lightning = "^2.2.2"
+python-telegram-bot = "^21.1.1"
 requests = "^2.28.2"
-torch = ">=1.11.0, !=2.0.1"
+torch = "^2.2.2"
 
 [tool.poetry.group.dev.dependencies]
-pytest = ">=7.3.1,<9.0.0"
-black = ">=23.3,<25.0"
-isort = {extras = ["colors"], version = "^5.11.4"}
-mypy = ">=0.991,<1.10"
-mypy-extensions = ">=0.4.3,<1.1.0"
-flake8 = "^6.0.0"
-safety = ">=2.3.5,<4.0.0"
-bandit = "^1.7.4"
-pre-commit = "^3.3.1"
-pytest-cov = ">=4,<6"
-coverage = "^7.2.5"
-flake8-docstrings = "^1.6.0"
-pyupgrade = "^3.3.1"
-ruff = ">=0.0.275,<0.4.1"
+ruff = "^0.4.0"
+safety = "^3.0.1"
+bandit = "^1.7.7"
+pre-commit = "^3.6.2"
+ipykernel = "^6.22.0"
+ipywidgets = "^8.0.6"
+ipython = "^8.23.0"
+jupytext = " ^1.16.1"
+nbconvert = "^7.16.3"
+
+[tool.poetry.group.test.dependencies]
+pytest = "^8.0.2"
+pytest-cov = "^5.0.0"
+pytest-mock = "^3.12.0"
+pre-commit = "^3.6.2"
+pytest-cases = "^3.8.1"
+
+[tool.poetry.group.docs.dependencies]
+mkdocs = "^1.5.3"
+mkdocs-material = "^9.5.18"
+mkdocs-git-revision-date-localized-plugin = "^1.2.4"
+mkdocs-awesome-pages-plugin = "^2.9.2"
+mkdocs-git-authors-plugin = "^0.8.0"
+pygments = "^2.17.2"
+mkdocs-jupyter = "^0.24.7"
+mkapi = "^2.1.0"
 
 [tool.poetry.group.debug]
 optional = true
 
 [tool.poetry.group.debug.dependencies]
 ipdb = "^0.13.9"
-line_profiler = "^4.0.3"
+line_profiler = "^4.1.2"
 
-[tool.poetry.urls]
-issues = "https://github.com/twsl/whos-there/issues"
+[tool.poetry.scripts]
+whos-there = "whos_there.cli:main"
+
+[tool.poetry-dynamic-versioning]
+enable = false
+vcs = "git"
+style = "semver"
+pattern = "default-unprefixed"
+
+[tool.poetry-dynamic-versioning.substitution]
+folders = [
+  { path = "src/whos_there" }
+]
+
+[tool.poetry-dynamic-versioning.files."src/whos_there/__about__.py"]
+persistent-substitution = false
+initial-content = """
+# file generated by poetry dynamic versioning during poetry build/install
+# don't change, don't track in version control
+__version__ = "0.0.0"
+__version_tuple__ = (0, 0, 0)
+"""
+
+[tool.pytest.ini_options]
+minversion = "8.0"
+addopts = "--doctest-glob='test_*.md'"
+testpaths = [
+    "tests",
+]
 
 [tool.coverage.run]
 branch = true
@@ -94,8 +135,8 @@ exclude_lines = [
 ]
 
 [tool.ruff]
-target-version = 'py310'
-select = [
+line-length = 120  # Must agree with Black
+lint.select = [
     "B",  # flake8-bugbear
     "C4", # flake8-comprehensions
     "D",  # pydocstyle
@@ -114,19 +155,15 @@ select = [
     "W",  # Warning
     "YTT", # flake8-2020
 ]
-
 exclude = [
     "migrations",
     "__pycache__",
-    "manage.py",
-    "settings.py",
     "env",
     ".env",
     "venv",
     ".venv",
 ]
-
-ignore = [
+lint.ignore = [
     "D100",
     "D101",
     "D102",
@@ -140,22 +177,22 @@ ignore = [
     "E402",
     "E501",
     "F401",
+    "TRY002", # Allow default exceptions
     "TRY003",  # Avoid specifying messages outside exception class; overly strict, especially for ValueError
 ]
-line-length = 120  # Must agree with Black
 
-[tool.ruff.flake8-bugbear]
+[tool.ruff.lint.flake8-bugbear]
 extend-immutable-calls = [
     "chr",
     "typer.Argument",
     "typer.Option",
 ]
 
-[tool.ruff.pydocstyle]
+[tool.ruff.lint.pydocstyle]
 convention = "google"
 
-[tool.ruff.per-file-ignores]
-"tests/*.py" = [
+[tool.ruff.lint.per-file-ignores]
+"test*/**/*.py" = [
     "D100",
     "D101",
     "D102",
@@ -170,69 +207,33 @@ convention = "google"
     "PGH001",  # use of "eval"
 ]
 
-[tool.ruff.pep8-naming]
+[tool.ruff.lint.pep8-naming]
 staticmethod-decorators = [
     "pydantic.validator",
     "pydantic.root_validator",
 ]
 
-[tool.ruff.isort]
-known-first-party = ["whos_there"]
+[tool.ruff.lint.isort]
 force-sort-within-sections = true
 
-[tool.black]
-line-length = 120
-target_version = ['py310']
-color = true
-exclude = '''
-/(
-    \.git
-    | \.hg
-    | \.mypy_cache
-    | \.tox
-    | \.venv
-    | _build
-    | buck-out
-    | build
-    | dist
-    | env
-    | venv
-)/
-'''
+[tool.ruff.format]
+docstring-code-format = true
+quote-style = "double"
 
-[tool.isort]
-py_version = 310
-line_length = 120
-profile = "black"
-include_trailing_comma = true
-multi_line_output = 3
-indent = 4
-lines_after_imports = -1
-force_sort_within_sections = true
-#color_output = true # causes issues with vs code formatting
+[tool.bandit.assert_used]
+skips = ['*_test.py', '*/test_*.py']
 
-[tool.mypy]
-# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
-python_version = 3.10
-pretty = true
-show_traceback = true
-color_output = true
+[tool.ruff.lint.pycodestyle]
+max-doc-length = 120
+max-line-length = 120
 
-allow_redefinition = false
-check_untyped_defs = true
-disallow_any_generics = true
-disallow_incomplete_defs = true
-ignore_missing_imports = true
-implicit_reexport = false
-no_implicit_optional = true
-show_column_numbers = true
-show_error_codes = true
-show_error_context = true
-strict_equality = true
-strict_optional = true
-warn_no_return = true
-warn_redundant_casts = true
-warn_return_any = true
-warn_unreachable = true
-warn_unused_configs = true
-warn_unused_ignores = true
+[tool.pyright]
+pythonVersion = "3.11"
+pythonPlatform = "Linux"
+include = ["src"]
+reportMissingImports = true
+reportMissingTypeStubs = false
+reportIncompatibleMethodOverride = true
+
+[virtualenvs]
+in-project = true
