+++ src/whos_there/__about__.py
@@ -1,0 +1,4 @@
+# file generated by poetry dynamic versioning during poetry build/install
+# don't change, don't track in version control
+__version__ = "0.0.0"
+__version_tuple__ = (0, 0, 0)
