+++ src/whos_there/callback.py
@@ -4,8 +4,9 @@ import socket
 import textwrap
 from typing import Any
 
-import pytorch_lightning as pl
-from pytorch_lightning.trainer.states import TrainerFn
+import lightning.pytorch as pl
+from lightning.pytorch.callbacks import Callback
+from lightning.pytorch.trainer.states import TrainerFn
 
 from whos_there.senders.base import Sender
 from whos_there.senders.debug import DebugSender
@@ -14,10 +15,10 @@ from whos_there.utils.logging import get_logger
 logger = get_logger(__name__)
 
 
-class NotificationCallback(pl.Callback):
+class NotificationCallback(Callback):
     """Notification callback."""
 
-    def __init__(self, senders: list[Sender] = None) -> None:
+    def __init__(self, senders: list[Sender] | None = None) -> None:
         """Initialize the notification callback.
 
         Args:
@@ -25,7 +26,7 @@ class NotificationCallback(pl.Callback):
         """
         super().__init__()
         self.senders: list[Sender] = senders if senders else [DebugSender()]
-        self._current_stage: str = None
+        self._current_stage: str = ""
 
     def _send(self, text: str) -> None:
         for sender in self.senders:
@@ -38,8 +39,8 @@ class NotificationCallback(pl.Callback):
         """Called when fit, validate, test, predict, or tune begins.
 
         Args:
-            trainer: The current :class:`~pytorch_lightning.trainer.Trainer` instance.
-            pl_module: The current :class:`~pytorch_lightning.core.lightning.LightningModule` instance.
+            trainer: The current :class:`~lightning.pytorch.trainer.Trainer` instance.
+            pl_module: The current :class:`~lightning.pytorch.core.lightning.LightningModule` instance.
             stage: The stage the trainer is currently in.
         """
         if trainer.global_rank == 0:
@@ -49,8 +50,8 @@ class NotificationCallback(pl.Callback):
         """Called when fit, validate, test, predict, or tune ends.
 
         Args:
-            trainer: The current :class:`~pytorch_lightning.trainer.Trainer` instance.
-            pl_module: The current :class:`~pytorch_lightning.core.lightning.LightningModule` instance.
+            trainer: The current :class:`~lightning.pytorch.trainer.Trainer` instance.
+            pl_module: The current :class:`~lightning.pytorch.core.lightning.LightningModule` instance.
             stage: The stage the trainer is currently in.
         """
         if trainer.global_rank == 0:
@@ -70,8 +71,8 @@ class NotificationCallback(pl.Callback):
         """Called when any trainer execution is interrupted by an exception.
 
         Args:
-            trainer: The current :class:`~pytorch_lightning.trainer.Trainer` instance.
-            pl_module: The current :class:`~pytorch_lightning.core.lightning.LightningModule` instance.
+            trainer: The current :class:`~lightning.pytorch.trainer.Trainer` instance.
+            pl_module: The current :class:`~lightning.pytorch.core.lightning.LightningModule` instance.
             exception: The exception raised.
         """
         name = pl_module._get_name()
@@ -94,8 +95,9 @@ class NotificationCallback(pl.Callback):
         r"""Called when loading a model checkpoint, use to reload state.
 
         Args:
-            trainer: the current :class:`~pytorch_lightning.trainer.Trainer` instance.
-            pl_module: the current :class:`~pytorch_lightning.core.module.LightningModule` instance.
+            trainer: the current :class:`~lightning.pytorch.trainer.Trainer` instance.
+            pl_module: the current :class:`~lightning.pytorch.core.module.LightningModule` instance.
             checkpoint: the full checkpoint dictionary that got loaded by the Trainer.
         """
-        self._current_stage = checkpoint.get("current_stage", None)
+        stage = checkpoint.get("current_stage")
+        self._current_stage = stage if stage else ""
