+++ src/whos_there/cli.py
@@ -1,0 +1,5 @@
+from whos_there.__about__ import __version__
+
+
+def main():
+    print(f"whos_there v{__version__}!")
