+++ src/whos_there/senders/slack.py
@@ -1,10 +1,12 @@
 from __future__ import annotations  # remove when dropping 3.8 support
 
+from typing import Any
+
 from whos_there.senders.base import Sender
 
 
 class SlackSender(Sender):
-    def __init__(self, webhook_url: str, channel: str, user_mentions: list[str] = None) -> None:
+    def __init__(self, webhook_url: str, channel: str, user_mentions: list[str] | None = None) -> None:
         """Initialize the Slack sender.
 
         Args:
@@ -17,7 +19,7 @@ class SlackSender(Sender):
         self.channel = channel
         self.user_mentions = " ".join(user_mentions) if user_mentions else []
 
-    def send(self, text: str) -> None:
+    def send(self, text: str) -> Any:
         data = {
             "username": "Knock Knock",
             "channel": self.channel,
