+++ src/whos_there/senders/teams.py
@@ -1,10 +1,12 @@
 from __future__ import annotations  # remove when dropping 3.8 support
 
+from typing import Any
+
 from whos_there.senders.base import Sender
 
 
 class TeamsSender(Sender):
-    def __init__(self, webhook_url: str, user_mentions: list[str] = None) -> None:
+    def __init__(self, webhook_url: str, user_mentions: list[str] | None = None) -> None:
         """Initialize the Teams sender.
 
         Args:
@@ -15,7 +17,7 @@ class TeamsSender(Sender):
         self.webhook_url = webhook_url
         self.user_mentions = " ".join(user_mentions) if user_mentions else []
 
-    def send(self, text: str) -> None:
+    def send(self, text: str) -> Any:
         data = {
             "username": "Knock Knock",
             "icon_emoji": ":clapper:",  # ":tada:"
