+++ src/whos_there/senders/telegram.py
@@ -1,4 +1,7 @@
+from typing import Any
+
 from telegram import Bot
+from telegram.constants import MessageLimit
 
 from whos_there.senders.base import Sender
 
@@ -14,16 +17,16 @@ class TelegramSender(Sender):
         super().__init__()
         self.token = token
         self.chat_id = chat_id
-        self._bot: Bot = None
+        self._bot: Bot | None = None
 
     @property
-    def bot(self):
+    def bot(self) -> Bot:
         if not self._bot:
             self._bot = Bot(token=self.token)
         return self._bot
 
-    def send(self, text: str) -> None:
-        length = 4096
-        chunks = [text[0 + i : length + i] for i in range(0, len(text), length)]
+    def send(self, text: str) -> Any:
+        length = MessageLimit.MAX_TEXT_LENGTH
+        chunks = [text[i : length + i] for i in range(0, len(text), length)]
         for chunk in chunks:
-            self.bot.send_message(chat_id=self.chat_id, text=chunk)
+            _ = self.bot.send_message(chat_id=self.chat_id, text=chunk)
