+++ src/whos_there/utils/logging.py
@@ -1,6 +1,6 @@
 import logging
 
-from pytorch_lightning.utilities import rank_zero_only
+from lightning.pytorch.utilities import rank_zero_only
 
 
 def get_logger(name=__name__) -> logging.Logger:
