+++ tests/unit/test_base.py
@@ -1,0 +1,5 @@
+from whos_there.__about__ import __version__
+
+
+def test_version():
+    assert __version__ is not None
